deleted graph data
